Move TargetActors to its own file
