+++ src/engine/mod.rs
@@ -1,20 +1,18 @@
 pub mod builder;
 pub mod incremental;
 mod target_actor;
+mod target_actors;
 mod watcher;
 
-use crate::domain::{Target, TargetId};
+use crate::domain::TargetId;
 use crate::TerminationMessage;
 use anyhow::{Context, Result};
 use async_std::prelude::*;
-use async_std::sync::{self, Receiver};
-use async_std::task::JoinHandle;
-use futures::{future, FutureExt};
-use std::collections::{HashMap, HashSet};
-use sync::Sender;
-use target_actor::{
-    ActorId, ActorInputMessage, ExecutionKind, TargetActorHandleSet, TargetActorOutputMessage,
-};
+use async_std::sync::Receiver;
+use futures::FutureExt;
+use std::collections::HashSet;
+use target_actor::{ActorId, ActorInputMessage, ExecutionKind, TargetActorOutputMessage};
+pub use target_actors::TargetActors;
 
 pub async fn run(
     root_target_ids: Vec<TargetId>,
@@ -129,88 +127,6 @@ async fn execute_once(
 
     Ok(())
 }
-
-pub struct TargetActors {
-    targets: HashMap<TargetId, Target>,
-    target_actor_output_sender: Sender<TargetActorOutputMessage>,
-    watch_option: WatchOption,
-    target_actor_handles: HashMap<TargetId, TargetActorHandleSet>,
-    target_actor_join_handles: Vec<JoinHandle<()>>,
-}
-
-impl TargetActors {
-    pub fn new(
-        targets: HashMap<TargetId, Target>,
-        target_actor_output_sender: Sender<TargetActorOutputMessage>,
-        watch_option: WatchOption,
-    ) -> Self {
-        let target_actor_handles = HashMap::new();
-        let target_actor_join_handles = Vec::new();
-
-        Self {
-            targets,
-            target_actor_output_sender,
-            watch_option,
-            target_actor_handles,
-            target_actor_join_handles,
-        }
-    }
-
-    fn get_target_actor_handles<'a>(
-        &'a mut self,
-        target_id: &TargetId,
-    ) -> Result<&'a TargetActorHandleSet> {
-        if !&self.target_actor_handles.contains_key(target_id) {
-            let (owned_target_id, target) = self.targets.remove_entry(target_id).unwrap();
-            let (join_handle, handles) = target_actor::launch_target_actor(
-                target,
-                self.watch_option,
-                self.target_actor_output_sender.clone(),
-            )?;
-            self.target_actor_handles.insert(owned_target_id, handles);
-            self.target_actor_join_handles.push(join_handle);
-        }
-
-        Ok(&self.target_actor_handles[target_id])
-    }
-
-    async fn send(&mut self, target_id: &TargetId, msg: ActorInputMessage) -> Result<()> {
-        self.get_target_actor_handles(target_id)?
-            .target_actor_input_sender
-            .send(msg)
-            .await;
-
-        Ok(())
-    }
-
-    async fn request_target(&mut self, target_id: &TargetId) -> Result<()> {
-        let handles = self.get_target_actor_handles(target_id)?;
-        for &kind in &[ExecutionKind::Build, ExecutionKind::Service] {
-            let build_msg = ActorInputMessage::Requested {
-                kind,
-                requester: ActorId::Root,
-            };
-            handles.target_actor_input_sender.send(build_msg).await;
-        }
-
-        Ok(())
-    }
-
-    pub async fn terminate(self) {
-        Self::send_termination_message(&self.target_actor_handles).await;
-        future::join_all(self.target_actor_join_handles).await;
-    }
-
-    async fn send_termination_message(
-        target_actor_handles: &HashMap<TargetId, TargetActorHandleSet>,
-    ) {
-        log::debug!("Terminating all targets");
-        for handles in target_actor_handles.values() {
-            handles.termination_sender.send(TerminationMessage).await;
-        }
-    }
-}
-
 #[derive(Copy, Clone)]
 pub enum WatchOption {
     Enabled,
